pyproject.toml
[tool.poetry]
name = "neural-style-transfer"
version = "0.1.0"
description = ""
authors = ["Kelvin Ducray <kelvin.ducray@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
black = "^21.12b0"
torch = "^1.10.1"
jupyter = "^1.0.0"
pydantic = "^1.9.0"
Pillow = "^9.0.0"
torchvision = "^0.11.2"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
mypy = "^0.930"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
